Fucking loads…
Made TITokenField responsible for a whole load of it's own delegate
stuff (still looking for a solution for the  shouldChange… delegate
method).

addToken: now accepts a TIToken. You can also use addTokenWithTitle:
which accepts an NSString

It now has a selectedToken property you can use to get the  selected
token and internally it uses it to deselect when another token is
selected, rather than looping over everything.

The add button and prompt label have been moved to leftView and
rightView properties rather than separate subviews.

TIToken is now a UIControl and doesn't have a delegate. You should
respond to events with addTarget:action:forControlEvents

It also has a property for a representedObject (like a custom Contact
class you might want to link to it).