Making sort of results optional
For some result sets we may not need the sorting of results. Making
it configurable so that caller can decide if it needs sorting. This
is particularly useful for default implementation as it performs
a brute force search on the source array which can be sorted once
by the caller and used thereafter as it is without cost of sorting
it everytime.