Allowing search in background thread
Allowing caller to set flag to search in background. This is
particularly useful when we have a long source array in which
we are going to search, which can potentially lead to blocking
of main thread for a long time.

Have tested it in the example by setting this flag to both YES
and NO. To maintain backward compatiblity just keeping the default
value to NO.