forbid auto-searching
Verbatim auto-searching always reault in UI stoping responding when
source array is very huge. It is always a good thing to allow user to
close the auto-searching feature although we can perform a background
search. Meanwhile should allow user to execute a manual search when
user need.This commit is just that.