feat(library): Added selected token color change for flat design
When a flat design token is selected / highlighted the color of the
selected state is made either slightly darker or slightly brighter.
This is determined based of whether or not the color is a 'dark' color
or if it is a 'light' color. A weighted distance of the RGB color space
is computed and if the color is greater than 127.5 it is darker.

Weighted RGB Brightness formula: brightness=sqrt( .241*R^2 + .691*G^2 + .068*B^2)

Added TIToken private methods:
- (UIColor *)lighterColorForColor:(UIColor *)c;
- (UIColor *)darkerColorForColor:(UIColor *)c;
- (BOOL)isColorDarker:(UIColor *)color;